requirements.txt
streamlit>=1.55.0
python-dotenv>=1.2.2
openai>=2.30.0
numpy>=1.26.0
